MB-68352: Simplify threads shutdown
As part of MB-66995 all clients are disconnected before we try to shut
down the front end threads. Previously we would schedule a task from
the thread doing the shut down (but not wait for it to run/complete)
which *could* cause multiple functions being scheduled, but not run
when we tell the event base to terminate.

As part of running the destructor for folly's event base it would
drain its notification queue and then run all functions which hasn't
been run yet. Given the order C++ runs destructors we ended up
referencing a destroyed object.

Given we know that we don't have any connected clients we can just
stop the threads and let then let the event pool drain the (potential)
events *before* releasing the resources.

Change-Id: I4f6bc11d48f8d33031df89e4fe34bb9890fe22d5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233750
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dan Owen <owend@couchbase.com>