MB-68518: Allow extra logging of responses from ext auth manager
When developing / debugging features involving external auth it is
often useful to see the reply from ns_server (as the error message may
be triggered internally etc).

With this patch an administrator may enable / disable logging by using
an ioctl value. Given that this isn't something one should use in
production it shouldn't require support in ns_server. Instead
this is controled with mcctl binary using the key
external.auth.logging

It may be set with:

    ./mcctl [options] set external.auth.logging true/false

And fetched with

    ./mcctl [options] get external.auth.logging

Change-Id: I82b01cfadac1ee099da1fe3057cfc88627a00966
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225498
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>