MB-68655: Test more than CAS when considering Item/SV equality
In the processExpiredItem path there is a StoredValue vs Item
compare that considers same CAS being same Item. This is not
correct - SetWithMeta could create a new mutation that differs
from what Item in the expiry pager is holding onto - but the
CAS could match.

This commit ensures all metadata is compared (and seqno) when
considering if an Item matches the StoredValue.

Change-Id: I7fa20a4b9156cf12ad3c0e7c782e1751994ed735
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234233
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>