MB-68594: [BP] Make gs2-authzid optional for OAUTHBEARER
Previously we required the gs2 header to specify no channel binding
and mandatory gs2-authzid.

This patch allows gs2-authzid to be optional and we'll pick out
the username from the returned RBAC entry from ns_server.

It is expected that *ns_server* honors the gs2-authzid as kv *WILL*
use that as the username if that's present in the request. If no
gs2-authzid is specified we'll pick the "one and only" username
specified in the returned "rbac" entry in the returned token
metadata:

In the returned payload below we would use "myusername" as the
user:

    {
      "token": {
        "exp": 1758705142,
        "rbac": {
          "myusername": {
            "buckets": {
              "default": {
                "privileges": [
                  "Read"
                ]
              }
            },
            "domain": "external",
            "privileges": []
          }
        }
      }
    }

Change-Id: Ie3ca4efd1df04ac59f962f5718750c295a6ae1cc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/238364
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>