MB-65251: Correctly update HCS in PDM for Disk Checkpoints
Committed items on disk  are sent as regular mutations via DCP,
during a backfill from disk. Therefore the HCS in the PDM is never
updated on the replica via the setMeta path.

The HCS for a disk snapshot is packed in the snapshot marker
(starting v2.0). When a disk snapshot is fully processed on the replica,
we can safely update the HCS in the PDM.

Change-Id: I2577ef7f29e9df09e364a2ce5576e5ecf931d16d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234418
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>