MB-68823: Don't throw EPEngine exception to core in StreamReq path
The exception object would be allocated in the EPE domain and then
wrongly released in the NoBucketDomain otherwise.

All frontend code paths into EPEngine suffer by that potentially (if
any path throws). A wider refactor for addressing that is in progress
on the master branch.

Note that the extra logging introduced in this patch replaces the
connection-details logging (provided at core level) that we lose by
catching the exception in EPE.

Change-Id: Ifacb6fb9970575c2177bbd3d26d85df226dee3e3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234445
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>