MB-61655: Add compat_version parameter
This parameter will hold the bucket configuration compatiblity version.

It will be initialised and updated by ns_server, to the cluster compat
version.

Defining it as a "normal" parameter in configuration.json, has the
benefit that the parameter is automatically reported in STAT(config),
which ns_server monitor for configuration differences between it's
expectation and actual values set in the bucket.

Change-Id: I339bd5f820358bb5825dd2a2e9e4e0bc2eb95f07
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234660
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>