MB-68753: Ensure new DcpConsumer is eligible for notifications
The DcpConsumer is contructed and leaving the parent member paused
as false. This means that ConnManager::run will ignore the connection
and that any local scheduleNotify will be a no-op.

This issue is behind the intermitent failure seen in MB-67544

This appears to work because we maybe relying upon step getting many
callbacks, and in the case where the connection uses its timeslice
no more callbacks until a notification is requested (which never
happens because of the paused=false issue).

Change-Id: I82c4b8045eca358ec9534962facacb966c487328
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234556
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Well-Formed: Restriction Checker