MB-64001: Re-enable CacheTransfer
We need to at least for now run the CacheTransfer on takeover
only.

This occurs because ns_server appears to do the follow.

1) Move snapshot from A to B
2) On B create a PassiveStream (which triggered cache)
3) On B immediately close stream and create
   a PassiveStream(takeover=true)

It appears this happens because the vbucket state on B appears
to have progressed enough for takeover (the stream at step 2
seems pointless and a ticket will track that).

However a CTS stream on A has dumped a bunch of cached values
onto the network, which then get processed by the takeover
stream and that fails due to opaque mismatch.

At least for now, and maybe forever just do CTS on takeover.

Change-Id: I54a4d9d7bc93f469168a5d3d2f03c2ec69c57d9e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234437
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>