MB-61655: Allow versioned parameter defaults to be specified
A map of versioned defaults can be specified:

  "default": {
    "7.2": 100,
    "8.0": 1000
  }

When the compat version is <8.0, and if the parameter is using defaults,
the parameter value will be updated to 100.

When the compat version is >=8.0, and if the parameter is using
defaults, the parameter value will be updated to 1000.

This defaults changing logic is implemented in a follow up patch. The
current patch only deals with storing the versioned map.

Change-Id: I2183fdf4a7e6ecf61d7a62360404ea620108cf9a

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234661
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>