MB-61655: Update defaults when compat_version changes
We should update the values for parameters using defaults when the
compat version changes.

We can do this in the processCompatVersionChange method, which is called
when a new compat version has been set.

We can track the set of explicitly set parameters, called "configured
parameters" here, to differentiate between parameters using defaults and
parameters using values equal to the defaults but explicitly set by
ns_server via the bucket config string or via the REST API (which calls
SetParameter).

Additionally, we track the alises of parameter names. We build a map,
which is only mutated during the Configuration object initialization,
and is only read thereafter, therefore it requires no synchronization.

Then, in processCompatVersionChange, we can go over all parameters, and
update the value of any which have not been configured, to the new
default as set by the defaults map and version.

Change-Id: Ie9f9e0689e996656e164638869764c185155e8bb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234662
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>