MB-69146: Add "topkeys-trace" functionality
When the cluster is one may collect a sample of the topkeys
with the following command:

     mctopkeys --host one-node-in-the-cluster \
               --cluster \
               --user Administrator \
               --password - \
	       --duration 10 \
	       --limit 1000 \
	       --decode-collections \
	       --collect-limit 10000

The command connects to the provided host (or localhost if no host is
specified) and fetch the global cluster map and connect to each node in
the cluster and start collecting usage of up to 'collect-limit' unique
keys for each front end thread (access to other keys is reported as
'overflow'). (Without --cluster it'll operate on the node specified)

When trace is disabled the overhead is "next to nothing" (an extra
check on a thread local variable). When enabled there is an extra
memory cost and cpu cost in order to look up the key). If there is
a "hot key" there might be mutex contention trying to access that
key. Given that it is intended to be used for "short samples" of
the current workload it may be an acceptable price to pay.

Change-Id: I989b465c079fcb6782152e926cb44b24499e701e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235905
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>