MB-68750: Add stat is_mouted to check if a fusion vbucket is mounted
Allow ns_server to query if vbuckets are mounted or not, then act
accordingly, e.g. perform UnmountVB on the ones that are in the mounted state after a partial rebalance.

Usage: Perform a stat call with the key in the following manner:
    "fusion is_mounted <vbid>"

Returns a Json value that consists of a key "is_mounted" and a boolean value. e.g.
    {"is_mounted", true}

Change-Id: I823aa075e153a54c210559e2396af07c57527268
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>