MB-67265: Replace check for fusionLogstoreURI with ...
... isFusionSupportEnabled

We have recently allowed initialising MagmaKVStore before the
logstore URI is set. Therefore, it is possible that fusionStats
are updated before the URI is set. This patch replaces the check
for fusionLogstoreURI with isFusionSupportIsEnabled().

Along with this change, we can remove the fusion stat test from
testapp_fusion.cc and add it to ep_testsuite.cc, such that all
stats are in a single place.

Change-Id: I2e65d82e53c903e081a4dc8ef843ccbb44b21fed
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235270
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>