MB-67175: Add termination reason to log description
Extend the description with the termination reason if it is
set:

    2025-10-16T11:13:31.741832+02:00 INFO Removing connection
    {"conn_id":925,
     "bucket":"travel-sample",
     "dcp":"producer",
     "dcp_name":"eq_dcpq:DCPT:INIT_309d2c13b63794168603c93e5ee53d84:travel-sample:inventory:airline:11100504861856972217:##4:3",
     "description":{
         "peer":{
             "ip":"127.0.0.1",
             "port":44664},
         "socket":{
             "ip":"127.0.0.1",
             "port":11210},
         "termination_reason":"Client closed connection: EOF",
         "user":{
             "name":"@projector",
             "system":true}}}

Previously the description was set when the connection was created and
only updated as part of authentication (and in those contexts it
wouldn't be used from other threads). DCP connections will log the
description from other threads than the connections front end thread
so with this change we need to protect the data with a lock.

Change-Id: I2fd4d730bbf55d42696d994fa3ba89d1422299c4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235358
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>