MB-64001: Add a new DCP opcode for transferring just key+meta
Adding DcpCachedKeyMeta 0x67

The cache-transfer protocol needs updating so that it can safely send a
key+meta (no value). The current CachedValue message cannot be used for
this task. The problem is with common validation code that is failing a
CachedValue (with no value) when the datatype is set. For example the
snappy validation will be triggered by the CachedValue handler and
fails because the 0 length value triggers an error. Manipulating the
common validation to handle the special case of the cache transfer is
not desirable, hence the addition of a new message which stores no
value and can have safe validation.

The majority of this patch (large) is boilerplate adding the
CachedKeyMeta through the engine/DCP APIs.

Note that CachedValue and CachedKeyMeta share the same format as
Mutation, only the opcode is different (and the old extended meta
formatting is never used).

Note that in this patch level, the validation has not been adjusted and
the issue of replicating the non-resident key/meta with !raw datatype
still exists (added in a separate commit).

Change-Id: I7320b63ae228f2464a00bfc80cf5dc0c03315b46
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235661
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>