MB-69046: Fix sporadic unit test failure in OAUTHBEARER
We've seen test failures for the OAUTHBEARER unit tests where the
token expire causing “auth stale” to be returned. The return code is
expected to be returned once the token expire, but in the test case it
isn’t expected to expire yet. The expiry time was set to the current
system time + 2 seconds, but that doesn’t necessarily mean we have 2
full seconds. It could be closer to 1 second if the system clock
passed the second barrier the next nanosecond.

We cannot really predict the speed of our CV builders as they run on
virtualized hardware, and there may be other tasks running in the
background. In addition to that when running on sanitizers things may
run extra slow.

This patch increase the timeout to be '4-5' seconds when running on
our normal builders (which *should* be enough even if the machine is
a bit slow), and 15 seconds when running under thread sanitizer.

This have the sad side effect that the test runtime increase with
the same amount of time, but its a lot faster than having to rerun
the entire build (or having to validate why there was a sporadic
unit test failure). I have tried to mock around with the command
to set the system time on the server, but that had the sad side
effect that all DCP streams got upset.

Change-Id: I1d20bd336cdd08a5642ed363ea8ee0e4f2365592
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235545
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>