MB-64001: Refactoring of DcpResponse/MutationResponse
The motivation of this first of many refactors is to help integrate
more of the CacheTransfer messages, primarily around the complex
constructor for MutationResponse (which influeces
MutationConsumerResponse)

It was noted that MutationResponse has 6 "Include..." type parameters
that are passed in construction and stored for the entire message.

Some of those parameters have no actual use and some are used only
in getMessageSize...

If we calculate the message size during construction we can ultimatley
remove the "long term" storage of those 6 parameters and reduce the
constructor parameters (a later patch).

For MutationConsumerResponse we can also remove the emd storage as it is
never used. 4.5 was the last build which could transmit
ExtendedMetaData. In the current code base, KVBucket::setWithMeta may
accept an emd parameter, but it is not passed down to
VBucket::setWithMeta... so we can now begin to remove that (and later
delete MutationConsumerResponse).

Note that an initial motivator for this change was to hopefully reduce
the size of MutationResponse, as that is a commonly allocated object
and user of readyQ memory. The final patches which do remove the
various "Include" member variables doesn't reduce the size, but the
overall clean-up seems useful and frees space/padding for legit growth
in future that may not increase the object size.

Change-Id: I1fc54f5e174de4ff7c02d985285a89676b81a6ea
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235761
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>