MB-69075: Protect BackfillManager shared_ptr with shared_mutex
The pointer may be reset concurrently in DcpProducer::closeAllStreams()
so a mutex is needed for copying. There was a mutex before, but it was
removed in MB-37520 for performance. While closeAllStreamsLock
serializes reset, it is not enough for safe access. A shared mutex is
expected to cause less contention.

Also added null checks where missing.

Change-Id: Ib5fa78f0d50549f2b1a984d0ceade0fd7ea39faf
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235766
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker