MB-67589: Fix MSVC warning C4456...
From the description:

The declaration of identifier in the local scope hides the declaration
of the previous local declaration of the same name. This warning lets
you know that references to identifier in the local scope resolve to
the locally declared version, not the previous local, which may or may
not be your intent. To fix this issue, we recommend you give local
variables names that do not conflict with other local names.

Change-Id: I9888dcdda4473c5c0355a1dabf5c80e6b244094c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235912
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>