MB-64001: Add a new DCP opcode for signalling end of cache transfer
Adding DcpCacheTransferEnd 0x68

To provide synchronisation with ns_server we need a way to signal
that a FBR/Fusion vbucket is ready for next steps of rebalance,
i.e. move to takeover or for non-takeover the vbucket is ready for
use (and safe for failover). This is critical for value-eviction.

This commit adds all the end to end plumbing to send and receive
the message - but no code yet sends or actually handles that message,
this is mostly boiler plate code.

Change-Id: Ib35d7e42ec8915f02d09a4382b7819f59befdec5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236267
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>