MB-64001: Trigger CacheTransfer once for snapshot created vbuckets
Remove the trigger code which is using stats and instead trigger
only for replica vbuckets that were created with "use_snapshot"
feature.

Add to Vbucket a (new) dedicated member for determining how a
vbucket was created (lots of churn to ensure makeVbucket can
set this).

Warmup, SetVbucket (non snapshot), FBR or Fusion.

Add to EPVbucket a flag to ensure CTS happens no more than once and
a function shouldUseDcpCacheTransfer that returns if cache-transfer
can be used.

We only want to do a cache transfer for replica vbuckets created
with use_snapshot=fbr|fusion, and only once and never after a
state transition.

* vbuckets created as !replica never transfer cache.
* vbuckets created from warmup or set-vb-state (not use_snapshot)
  never transfer cache.
* vbuckets that change state disable cache transfer. E.g. don't allow
  a replica with use_snapshot=fbr to flip to active and back again.

Change-Id: I263dd8a1c985beaf3f155cfefec0d0a7d1d75b52
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/235996
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>