MB-69134: Change Warmup to work with less access logs than shards
Prior changes to the access scanner mean that there may not always
be an access log file for a shard - in the case that there is no
vbucket assigned/mapped to the shard there will be no file.

Warmup needs to handle this case and change away from using the shard
count to guide task creation and state transition.

Change-Id: Ia0a5a8261f5685be57f4051e2ca93773eb4f3ae2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236036
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker