MB-64001: Add CacheTransfer "all_keys" for value-eviction
The "all_keys" mode of CacheTransfer is the current solution for
providing value-eviction correctness in FBR (and even Fusion).

Since value eviction requires at least all of the keys to correctly
operate, the "all_keys" option for CacheTransfer will provide that.

The behaviour is:

1) Value eviction vbucket requests "all_keys":true
2) And also sets "free_memory" (which can be 0)

The cache transfer task will now visit all hash-table items and
(remember that availableBytes begins as the value provided by the
consumer and is decremented as the producer sends items).

1) Send a CacheValue if the item is resident and availableBytes > 0
2) Send a CacheKeyMeta if the item is not resident
3) Send a CacheKeyMeta when availableBytes==0 even for resident items

I.e. At the start of transfer we 1 or 2 applies, but as soon as the
availabeBytes is 0 then only 3 applies.

This ensures the replica receives all keys and ideally maintains a
resonable residency.

Change-Id: I81c026c72f85ffec470b65160f441b841c8a0e58
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236161
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>