MB-69328: Add FBR/Fusion stat to interlock with cache transfer
Provide a new stat group and stat to return if the cache transfer
has completed and the vbucket move can continue.

vb_1:snapshot_rebalance_continue true

The expectation is that ns_server is only reading/using this stat
after a vbucket was created from a snapshot. E.g. when a setvbucket
used the "use_snapshot=fbr|fusion" option. In those cases the
vbucket is expected to have an empty cache and require cache
population (network cache transfer). The flag remains false until
KV concludes the transfer (CacheTransferEnd received). ns_server
will wait on this stat in the fusion/fbr vbucket move path before
proceeding to takover or before concluding a replica move is
ready.

If a fbr/fusion vbucket changes state we also flip this flag to
true.

For any vbucket created by warmup or setvbucket then this value
defaults to true.

Change-Id: I66354b4f44d49b543612989497613328ab8ccdfe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236359
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>