MB-68855: Refactor fusion stat detail/aggregate
This patch introduces a new pattern for fusion stats, where:

  1. 'fusion <stat_group> <vbid>' returns the <stat_group> information
    for the given <vbid>
  2. 'fusion <stat_group> detail' returns the <stat_group> information
    for all vbuckets (for the bucket passed in the STAT call)
  3. 'fusion <stat_group>' returns the aggregated <stat_group>
    information from all vbuckets. Entries in the aggregated result are
    unique, see example below.

Eg for active_guest_volumes, let's assume that we have vb_A on volumeX
and vb_B again on volumeX, then:

  1. 'fusion active_guest_volumes vb_A' -> res: ["volumeX"]
  2. 'fusion active_guest_volumes detail'
       -> res: {"vb_A":["volumeX"], "vb_B":["volumeX"]}
  3. 'fusion active_guest_volumes' -> res: ["volumeX"]

Note that to date active_guest_volumes is the only fusion's stat group
that exposes the aggregate.
Note also that before this path the stat strings for (2) and (3) used
to behaves differently depending on the specific stat group. This patch
fixes that and consolidates teh format and semantic for all groups.

Change-Id: I2e81b0435e8fa90f068a3ec9793b3a97f3cbdb43
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236720
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>