MB-69238: Avoid exception for prepare keys
The document viewer in the UI requests keys by using the GetAllKeys
command, but that command may fail if the requested keys happens
to find a "prepare" in the included keys.

The code tries to ignore these keys, but tries to create a
DocKey before running the check, and the constructor of DocKey
throws the exception for Prepare keys.

Change-Id: Ib11749e5d6a5417c7219abe9b9e192152283fb0b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236613
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Well-Formed: Restriction Checker