MB-68465: Refactor get/release fusion storage snapshot to accept ...
... multiple vbuckets

These APIs internally do a read+write to Chronicle. Right now, ns_server
issues one call per vb, which translates to 30*1024 calls (in the worst
case of 30 buckets and 1024 vbs) and hence 30k small writes to Chronicle.
This is not the most efficient method for writing data to Chronicle.

With a bulk API, the writes will become much more efficient, and the
overall prepareRebalance API will speed up.

Change-Id: I6b346a991969eb3ab9901e71dd8c6123cfc8aac5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236802
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>