MB-64385: Set bucket throttling configurations
Extract the throttling bucket configs before they are passed to the
engine and set them within memcached bucket.

ValidateBucketConfig will ensure the value is a valid size_t and return
it as part of the validation json.
SetParam will extract the key and update it within memcached core and
not pass it to ep-engine.

Change-Id: Ic3735a0b586ba51b4d189188b0b984e751114ec5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236805
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>