MB-69554: Allow for tuning max connection dynamically
Previously one could only *reduce* the maximum number of connections
as the rest of the available file descriptors was given to the
underlying engines to use.

This patch change the logic we distributed the available file
descriptors by using a "minimum" configuration of:

1. 5k system connections
2. 60k user connections
3. 2k reserved for the core
4. 30k reserved for ep-engine
5. 2k for magma

If the available file descriptor limit is lower than the above
it does a "best effort" to divide the available file descriptors
into various pools (and disables the ability for the operator
to change connection counts)

Anything *above* 99000 may be allocated to connections by tuning the
configuration parameter max_connections. The rest of the file
descriptors will be given to magma. As part of *changing* the maximum
number of connections magma is told to reduce its file descriptor
limit (and close as many files as it needs to in order to get below
the new limit)

During startup memcached will log information like:

    INFO System file descriptor settings { \
         "connection_limits": \
	   {"system":5000,"user":65000}, \
	 "magma": \
	   {"current_open":0,"max_limit":903000}, \
	 "max_fds":1000000, \
	 "reserved":{"core":2000,"ep_engine":30000}}

If the available file descriptors is lower than the "expected minimum"
causing us to try to do a "best effort" divide the file descriptors
this log message will change its wording to reflect that:

    WARNING System file descriptor setting too low to support \
    desired number of connections; adjusted settings { \
      "connection_limits":{"system":3000,"user":31500}, \
      "magma":{"current_open":0,"max_limit":1500}, \
      "max_fds":50000, \
      "reserved":{"core":2000,"ep_engine":15000}}

(The severity is CRITICAL if the available file descriptors
is less than 20k and ERROR if it is lower than 30k)

Change-Id: Ia36829de8f07d519519fc0ca65d5884615bf990d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236946
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>