MB-69513: Remove call to get error map as part consumer setup
This code was added in order to detect a pre 5.0 producer. Not only is
anything before 7.0 EOLd, but it is also impossible to upgrade
directly to Totoro from any of the "affected" versions as one would
need to upgrade via a newer version anyway.

This extra complexity should be removed to simplify the setup code
(and avoid an extra roundtrip to the producer)

Change-Id: If1a808516f282d22fdfdb1816487bef2207c2109
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/236890
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>