MB-69242: Add cb::snapshot::CRCSink
Create a CRCSink class that will be used in the receiving side of a
GetFileFragment + crc request.

The test code generates a "stream" of "file-data" + crc replicating the
format that GetFileFragment will generate when CRC is requested.

GetFileFragment is to be enhanced so that a checksum_length can be
provided, which states how many bytes should be checksummed.

The resulting stream written to the network would then look like:

0 ...                  file-size
[n-file-bytes][crc][n-file-bytes][crc]

Accounting for when the requested length doesn't evenly divide into
equal n-file-byte chunks.

Change-Id: I6117a9911ed4dda4ad16fd58ed5d9aeff862273c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237052
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>