MB-69242: Add optional checksumming to GetFileFragment
Add the following optional configuration to the JSON object that
configures GetFileFragment.

"checksum_length" : "bytes"

When this value exists and is not 0 GetFileFragment will checksum
the file and provide the checksums in the file network stream.

E.g. when the checksum_length is n the network stream will consist
of chunks of file n-bytes long with the checksum of the previous
n-bytes appended.

{n-file-bytes}{4-byte-crc}...{m-file-bytes}{4-byte-crc}

Depending on the file size or read length the m-file-bytes section
maybe smaller or equal to n (the tail of the read).

Change-Id: I096d3aa99fa25d36097b524d1c0923910dcf9a5f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237053
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>