MB-69242: Make prepare_snapshot checksum optional
There is quite a cost in rebalance time if we generate
checksums for prepare snapshot, this is mostly due to
needing to read the entire file at the source during
the prepare and then again at the destination post
download.

The checksum in this process is though incredibly valuable
to the correctness of File-Based Rebalance, and as such
we should aim to run as much of our testing and validation
with this coarse checksumming enabled.

This patch adds mechanisms so that we can more easily run
with checksums disabled for testing that may needs to
validate performance.

1) When we build with CB_DEVELOPMENT_ASSERTS default to
   generating checksums.

2) A new environment variable can override this default

Setting the following in the environment of memcached will
disable checksums in prepare.
EP_PREPARE_SNAPSHOT_CHECKSUM_DISABLE

3) A config setting exists to override as a fail safe which
defaults to off/false.

Change-Id: I64b626e06c8e184ff72a134d3fcb305e7e96a7ff
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237099
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>