MB-69651: Limit the concurrency for DownloadSnapshotTask
The DownloadSnapshotTask connects to the remote server and performs
the following actions:

1. create a snapshot
2. download the entire snapshot locally

These steps may take quite some time and the task occupies one
thread in the AuxIO thread pool until the task is done. The tasks
it performs involves running tasks on the remote server on the
same thread pool, and as the observant reader already notice this
may easily end up in a deadlock (both parties are running the
DownloadSnapshotTask on all AuxIO threads requsting files from
each other and the other side do not have any available threads).

As a simple workaround for the problem until we come up with
a better solution this patch introduce a concurrency limit on
the number of downloads to the number of available AuxIO threads
- 2 (with a minimum on 1)

Change-Id: I7a0bb038a5facfbf1cbc71f4f877c633d80596a0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237349
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>