MB-68600: Add ImpersonateWithTokenAuthDataId frame id
Add the frame info identifier to represent the previously registered
authentication token. The size of the identifier should be 16 bits.

Given that the identifier should represent an authentication token
already registered with the server we may generate the privilege mask
"up front" as part of validating the command header and stop further
processing for unknown identifiers or if the token has expired.

Change-Id: Ibf6ccf9681144e3c74e0ec820b4901ce4261cb93
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237546
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>