MB-69678: Transition CTStream->ActiveStream by atomic-swap in streamMap
By https://review.couchbase.org/c/kv_engine/+/232769 we have introduced
a time window where a logically-open stream isn't present in the
Producer::streamMap when the stream tarnsitions from CacheTransfer to
ActiveStream.

In MB-69678 the symptom is that we fail ns_server's STAT(dcp-takeover)
calls by stream_not_found.

So this patch fixes that ns_server failure but in general it
re-introduces the invariant that a logically-open stream is always
referenced in Producer::streamMap. A stream is removed from the
streamMap only when it is logically closed.

Change-Id: I11153ede9114ec9a1fa76a55ea3186034498b43c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237610
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>