MB-64001: Construct keydata with the correct size
Noted whilst profiling on OSX/Clang that the hot StoredDocKey ctor was
spending 1.4% in basic_string::append - which was coming from the
resize call.

Changing the keydata ctor removed this and resulted in a more expected
profile where we see just memset in use.

I do note though it looks like (possibly) both bzero and memset (to set
the default value) are used, when really we don't care as the memory
will be overwritten - hinting that we maybe should just use
std::vector - but we maybe benefitting from small string optimisation. So for now leaving as std::string and removing the unexpected append
cost.

Change-Id: Ie9cadf5a3c795555f2cb49a215af58b611b1df40
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237618
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>