MB-69242: More efficient construction of keys on various hot paths
StoredValue::toItem and DCP ingestion both show from profiling
unnecessary cost in copying keys, and this stems from the way The
current key types have been spread throughout our code.

History lesson:

7.0 collections standardised the keys (no strings or char*/len).

1. StoredDocKey is an owning type - it use a std::string for storage.
The stored key in this case is always a “collection prefixed” key.

2. SerialisedDocKey is exclusively for HashTable use and uses the same
memory as the StoredValue. The key again is always “collection
prefixed”.

3. DocKeyView is a non-owning type which exists to cheaply pass/view
either “collection prefixed” keys or “legacy” keys, ones with no leb128
collection prefix. We would only actually encounter a non-prefixed key
in frontend paths for storing or reading documents and only when the
connection has not enabled collections.

However in general we use DocKeyView pretty much everywhere to pass/view
any key and in the bulk of these cases, the key is always a “collection
prefixed” key. This leads to minor inefficiencies in hot code paths
that have to create StoredDocKey from DocKeyView.

This patch introduces a new non-owning key-type which ultimately should
replace the majority of DocKeyView (or switch them around at some
point).

PrefixedDocKeyView is a non-owning key-type that can only
view “collection prefixed” keys. When this type is used the creation of
a StoredDocKey is now more optimal in that a single alloc/copy can be
used.

This patch then changes two paths where it was noticed in CPU profiling
that the current StoredDocKey construction was slightly more costly
than it could be.

StoredValue::toItem is frequently invoked in FBR/Fusion rebalance by the
new CacheTransferTask as eligible items are copied over to the DCP
readyQ.

Additionally on the consumer side, creation of Item objects on the DCP
ingestion path showed similar copying costs.

Both these cases the DocKeyView is in always viewing a “collection
prefixed” key, meaning we can use a leaner construction path (pass a
PrefixedDocKeyView).

A second minor change in Item construction is to call
isInSystemEventCollection on the StoredDocKey, not the DocKeyView. The
StoredDocKey variant is fractionally more efficient than DocKeyView as
there is no branching needed the encoding(a very minor cost appearing
in profiling of DCP ingestion).

Change-Id: I1d5c0d0c6b73aebf9da8b9a7173677ffce8e2454
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237838
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>